Splitting APIClient into definition and extensions for each model you can request
